udp-encapsulation-tunnel: remove unused inner IP+UDP layers
Before this patch, the UDP packet was containing in its data the IP
header + an extra UDP layer. On the receive side only the IP's ToS was
used from this inner layer, which is currently not used (so 0).

These two layers can then be dropped. So instead of having: IPv4 + UDP +
[IPv4 + UDP + TCP + payload], we have IPv4 + UDP + [TCP + payload]. In
other words, only the TCP and the data inside the UDP tunnel, that's all
we need here.

While at it, 'udp_checksum()' is then no longer needed. Same for the UDP
header info.

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>